+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 -   z_index from `Group`
 -   Misc. unused draw, setup, and update methods for some classes.
 -   Advanced rotational physics. Will be added in a later patch once Hitbox is refactored.
+-   Vector `unit()` method. Use `normalized()` or `normalize()` instead.
 
 ### Fixed